### 2022-02-12 #4155 @jbytecode
@kthyng - I feel comfortable with R but I have nothing with the cost-effectiveness analysis and health economic modeling.
