### 2022-02-25 #4206 @editorialbot
[REVIEW]: BCEA: An R Package for Cost-Effectiveness Analysis
**Submitting author:** @n8thangreen (<a href="http://orcid.org/0000-0003-2745-1736">Nathan Green</a>)
**Repository:** https://github.com/n8thangreen/BCEA/
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v2.4.2
