### 2022-03-09 n8thangreen/BCEA#93 @rowlandseymour
Contributor guidelines
One requirement for JOSS is that you have contributor guidelines. I can't find any currently for this package, although you do have a code of conduct. You clearly use the issues feature, so will defer to the the editor here. If you would like to see an example of contributing guidelines, you can find one from a recent JOSS paper here https://github.com/ejikeugba/serp/blob/43049d3c8fd55f51952e5354ccaa4997383442d0/contributing.md

openjournals/joss-reviews#4206
