### 2022-03-09 n8thangreen/BCEA#94 @rowlandseymour
Travis CI
It's good to see use of CI to check the package, but you're currently using Travis. [rOpenSci recommend moving away from Travis ](https://ropensci.org/blog/2020/11/19/moving-away-travis/)to GitHub Actions, but this is a completely personal choice and won't affect my review.

https://github.com/openjournals/joss-reviews/issues/4206
