### 2022-03-09 n8thangreen/BCEA#95 @rowlandseymour
JOSS paper
The paper reads well and gives a good summary of the package and why it's needed. Two very minor comments here:
* The formatting of the corresponding author has gone a bit awry,
* On the fourth line of the statement of need, the acronym PSA is not defined.
https://github.com/openjournals/joss-reviews/issues/4206
