### 2022-03-09 n8thangreen/BCEA#97 @rowlandseymour
Good practice suggestions
Running `goodpactice::gp()` gives the following suggestions. I don't consider any of these major, but you may want to consider them.
```
── GP BCEA ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write short and simple functions. These functions have high cyclomatic complexity:evppi.bcea (151).
  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug tracker. Many online code hosting services provide bug
    trackers for free, https://github.com, https://gitlab.com, etc.
  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users and developers are used it and it is easier to
    read your code for them if you use '<-'.

    R/bf.R:29:34
    R/bf.R:54:93
    R/bf.R:100:93
    R/eib_plot_graph.R:252:17
    R/evi_plot_graph.R:156:15
    ... and 24 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters

    R/base_methods.R:7:1
    R/base_methods.R:16:1
    R/base_methods.R:17:1
    R/bcea.R:12:1
    R/bcea.R:16:1
    ... and 407 more lines

  ✖ omit trailing semicolons from code lines. They are not needed and most R coding standards forbid them

    R/bf.R:29:44
    R/bf.R:33:30
    R/bf.R:34:67
    R/bf.R:39:43
    R/bf.R:42:42
    ... and 94 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using
    vapply() instead.

    R/bf.R:35:31
    R/ceac_plot_graph.R:166:7
    R/ceac_plot_graph.R:169:7
    R/ceac_plot_graph.R:178:5
    R/ceplane_plot_graph.R:215:16
    ... and 7 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions. They are error prone and result
    1:0 if the expression on the right hand side is zero. Use seq_len() or seq_along() instead.

    R/CreateInputs.R:154:10
    R/evppi_helpers.R:478:38
    R/lad.R:201:15
    R/lad.R:229:17
    R/lad.R:243:17
    ... and 2 more lines
```

https://github.com/openjournals/joss-reviews/issues/4206
