### 2022-03-09 n8thangreen/BCEA#98 @rowlandseymour
Vignette signposting
There are quite a few vignettes and I think there needs to be some signposting for the user when they arrive at the package about what vignettes to look at first or which vignettes to use for particular tasks. This could be in the README or elsewhere. Some of the vignettes appear to be incomplete and I think these either need completing or removing from the JOSS branch of the repository.

https://github.com/openjournals/joss-reviews/issues/4206
