### 2022-03-09 #4206 @rowlandseymour
I've completed my review of the package and overall I think this is a very well executed package and submission for JOSS. The package is already on CRAN and has a user base meaning that you meet a high standard for usability and contribution. There is a book on using the package, which is the ultimate form of documentation. Running goodpractice and dev tools checks highlight only very minor concerns, showing how well the package is written. I was able to run the functions and generate output and plots.

I have raised five issues, none of which are major, two are minor and the rest I would consider at the authors discretion. The first minor issue is about guiding the user towards vignettes as I don't think it is obvious where they should begin and some of the vignettes are very sparse. The second issue concerns contributor guidelines and I don't think they meet JOSS' current standards.

Thanks to @n8thangreen and the other authors, you made this very straightforward to review.
