### 2022-03-10 1d5c281
add JOSS-paper GitHub Action
