### 2022-03-10 daf593c
small edit fixes of JOSS paper closes #95
