### 2022-03-16 #4206 @n8thangreen
I believe I have now addressed all of the review comments. Thanks @rowlandseymour for the time and effort to review.
I have closed the issues in the BCEA repo corresponding to the JOSS review.

What are the next steps, please? Is there something else that I need to now do?
Thanks!
