### 2022-03-16 #4206 @rowlandseymour
Thanks @n8thangreen, I think the table of contents in the vignette is an excellent edition.

@ajstewartlang All the items on my checklist are now ticked off.
