### 2022-07-15 n8thangreen/BCEA#110 @MikeLydeamore
Data names override default R functions
There are a few named objects in the package datasets that are overwriting default R objects.

In Smoking: c and pi override the function and constant.
In Vaccine, c overrides the concatenate function.

In both cases, the bracketing means that the function and the dataset are accessible at the same time, but I could imagine an edge case (probably in `do.call`) that would cause some unexpected behaviour.

Could these be renamed so that if the data is loaded there can't be an accidental mixup?

https://github.com/openjournals/joss-reviews/issues/4206
