### 2022-07-15 #4206 @MikeLydeamore
I have completed my review of the package. Overall I agree with the previous reviewer. The package is very thoroughly documented, works as expected and has good development practices.

I believe it to be a valuable addition to the journal.

My only recommendation is the naming of a few of the included datasets be renamed to avoid overriding R functions with datasets.
