### 2022-07-25 #4206 @n8thangreen
I've changed the e, c and pi objects in the package supplied data to eff, cost and pi_post. I've pushed to the default branch dev. Is this sufficient? Thanks.
