### 2022-07-28 0e57c20
names in reference to journal style
closes #116
