### 2022-07-28 5469ad1
remove unused references
closes #114
