+++ paper.bib
@@ -17,26 +17,6 @@ url = {https://www.springer.com/gp/book/9783319557168},
 year = {2017}
 }
 
-@book{GoF,
-  added-at = {2010-06-05T16:40:25.000+0200},
-  asin = {0201633612},
-  author = {Gamma, E., Helm, R., Johnson, R., & Vlissides, J. M.},
-  biburl = {https://www.bibsonomy.org/bibtex/27e3f1154ab1fbce54752a46dba7f2217/pnk},
-  description = {Amazon.com: Design Patterns: Elements of Reusable Object-Oriented Software (9780201633610): Erich Gamma, Richard Helm, Ralph Johnson, John M. Vlissides: Books},
-  dewey = {005.12},
-  ean = {9780201633610},
-  edition = 1,
-  interhash = {7fe32957be97afaf4ecb38b5490d23b4},
-  intrahash = {7e3f1154ab1fbce54752a46dba7f2217},
-  isbn = {0201633612},
-  keywords = {DBIS Design Object-Oriented Patterns SS2010 Seminar Software},
-  publisher = {Addison-Wesley Professional},
-  timestamp = {2010-06-05T16:40:25.000+0200},
-  title = {Design Patterns: Elements of Reusable Object-Oriented Software},
-  url = {http://www.amazon.com/Design-Patterns-Elements-Reusable-Object-Oriented/dp/0201633612/ref=ntt_at_ep_dpi_1},
-  year = 1994
-}
-
 @article{Incerti2021,
 abstract = {Health economic models simulate the costs and effects of health technologies for use in health technology assessment (HTA) to inform efficient use of scarce resources. Models have historically been developed using spreadsheet software (e.g., Microsoft Excel) and while use of R is growing, general purpose modeling software is still limited. hesim is an R package that helps fill this gap by facilitating parameterization, simulation, and analysis of economic models in an integrated manner. Supported model types include cohort discrete time state transition models (cDTSTMs), individual continuous time state transition models (iCTSTMs), and partitioned survival models (PSMs), encompassing Markov (time-homogeneous and time-inhomogeneous) and semi-Markov processes. A modular design based on R6 and S3 classes allows users to combine separate submodels for disease progression, costs, and utility in a flexible way. Probabilistic sensitivity analysis (PSA) is used to propagate uncertainty in model parameters to model outputs. Simulation code is written in C++ so complex simulations such as those combining PSA and individual simulation can be run much more quickly than previously possible. Decision analysis within a cost-effectiveness framework is performed using simulated costs and quality-adjusted life years (QALYs) from a PSA.},
 archivePrefix = {arXiv},
@@ -65,8 +45,6 @@ volume = {1},
 year = {2017}
 }
 
-
-
 @article{Baio:2013,
 	author = {Baio, G.},
 	title = {{Bayesian models for cost-effectiveness analysis in the presence of structural zero costs}},
@@ -75,255 +53,6 @@ year = {2017}
 	volume = {33(11)},
 	pages = {1900-1913},
 }
-
-@book{Baio:2012,
-	author = {Baio, G.},
-	title = {{Bayesian Methods in Health Economics}},
-	publisher = {Chapman Hall/CRC Press, Boca Raton, FL},
-	year = 2012,
-}
-
-@book{Briggsetal:2006,
-	author = {Briggs, A., Sculpher, M., & Claxton, K.},
-	title = {{Decision modelling for health economic evaluation}},
-	publisher = {Oxford University Press, Oxford, UK},
-	year = 2006,
-}
-
-@article{Claxton:1999b,
-	author = {Claxton, K.},
[…]
