### 2022-07-28 d7f5d03
add dampack citation
closes #115
