+++ paper.bib
@@ -101,3 +101,12 @@ title = {{A Multidimensional Array Representation of State-Transition Model Dyna
 url = {http://biorxiv.org/content/early/2019/06/21/670612.abstract},
 year = {2019}
 }
+
+@Misc{dampack2021,
+  title = {Decision Analytic Modeling Package (dampack)},
+  author = {Fernando Alarid-Escudero and Greg Knowlton and Caleb
+    Easterly and Eva Enns},
+  year = {2021},
+  note = {R package version 1.0.0},
+  url = {https://github.com/DARTH-git/dampack},
+}
