+++ paper.md
@@ -39,7 +39,7 @@ The R (@R) package \texttt{BCEA} can be used to post-process the sampled costs a
 \texttt{BCEA} is a tool for interpreting and presenting the random sample of results from a CEA in a simple, powerful and standardised way, with useful, technically advanced measures and graphical summaries.
 \texttt{BCEA} was primarily written to use posterior distribution samples from a Bayesian model (e.g. run in WinBUGS or Stan) but can take any probability sensitivity analysis (PSA) random samples as inputs. \texttt{BCEA} also aims to be used in a health economic modelling workflow, meaning that it can be plugged-in as one of the steps in a CEA analysis.
 \texttt{BCEA} does not provide modelling functionality, like some other CEA packages such as \texttt{hesim} (@Incerti2021) or \texttt{heemod} (@Filipovic-Pierucci2017), but the package philosophy (borrowed from UNIX) is to do one thing well by focusing on the analysis following a model run.
-The package \texttt{dampack} is also a decision analytic modelling package which provides a suite of functions for analyzing and visualizing the health economic outputs but takes a different design approach to \texttt{BCEA}.
+The package \texttt{dampack} (@dampack2021) is also a decision analytic modelling package which provides a suite of functions for analyzing and visualizing the health economic outputs but takes a different design approach to \texttt{BCEA}.
 Although it has a focused scope, within this \texttt{BCEA} is designed to be extensible and flexible. Currently, \texttt{BCEA} has base R, \texttt{ggplot2} and \texttt{plotly} versions of the plotting functions.
 The code is written so that computation of new statistics and new plotting functionality can be easily added. In \texttt{BCEA} the workflow centres around the \texttt{bcea()} function rather than separate functions for each type of statistic, with the aim to reduce the learning curve and easily expose the package functionality.
 Finally, \texttt{BCEA} has an expansive suite of functions from basic cost-effectiveness analyses, e.g. increment benefit (IB) and ICER calculation and plotting, to more sophisticated methods, e.g. Expected Value of Perfect Partial Information (EVPPI).
