### 2022-07-28 #4206 @ajstewartlang
@openjournals/dev any idea why the `check references` command isn't working on this submission? The pdf builds ok and has the reference list in it. The `.bib` file looks ok to me.
