### 2022-08-08 #4206 @ajstewartlang
:wave: Thanks @n8thangreen this all looks good to me now - and almost ready to go. In the `pdf` there are some cases where the in-text citations don't look quite right.

For example, the following:
G. Baio (2013) and B. Baio G. (2017)

should be:
Baio (2013) and Baio (2017)

If you could take a look and try to fix the issue that would be great. In addition, could you also do the following please?

- Make a [tagged release](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository) of the software.
