+++ paper.bib
@@ -10,7 +10,7 @@
 
 @book{Baio2017,
 abstract = {The book provides a description of the process of health economic evaluation and modelling for cost-effectiveness analysis, particularly from the perspective of a Bayesian statistical approach. Some relevant theory and introductory concepts are presented using practical examples and two running case studies. The book also describes in detail how to perform health economic evaluations using the R package BCEA (Bayesian Cost-Effectiveness Analysis). BCEA can be used to post-process the results of a Bayesian cost-effectiveness model and perform advanced analyses producing standardised and highly customisable outputs. It presents all the features of the package, including its many functions and their practical application, as well as its user-friendly web interface. The book is a valuable resource for statisticians and practitioners working in the field of health economics wanting to simplify and standardise their workflow, for example in the preparation of dossiers in support of marketing authorisation, or academic and scientific publications.},
-author = {Baio, G., Berardi, A., & Heath, A.},
+author = {{Baio}, G. and {Berardi}, A. and {Heath}, A.},
 publisher = {Springer International Publishing},
 title = {{Bayesian Cost-Effectiveness Analysis with the R package BCEA}},
 url = {https://www.springer.com/gp/book/9783319557168},
@@ -21,7 +21,7 @@ year = {2017}
 abstract = {Health economic models simulate the costs and effects of health technologies for use in health technology assessment (HTA) to inform efficient use of scarce resources. Models have historically been developed using spreadsheet software (e.g., Microsoft Excel) and while use of R is growing, general purpose modeling software is still limited. hesim is an R package that helps fill this gap by facilitating parameterization, simulation, and analysis of economic models in an integrated manner. Supported model types include cohort discrete time state transition models (cDTSTMs), individual continuous time state transition models (iCTSTMs), and partitioned survival models (PSMs), encompassing Markov (time-homogeneous and time-inhomogeneous) and semi-Markov processes. A modular design based on R6 and S3 classes allows users to combine separate submodels for disease progression, costs, and utility in a flexible way. Probabilistic sensitivity analysis (PSA) is used to propagate uncertainty in model parameters to model outputs. Simulation code is written in C++ so complex simulations such as those combining PSA and individual simulation can be run much more quickly than previously possible. Decision analysis within a cost-effectiveness framework is performed using simulated costs and quality-adjusted life years (QALYs) from a PSA.},
 archivePrefix = {arXiv},
 arxivId = {2102.09437},
-author = {Incerti, D., & Jansen, J. P.},
+author = {{Incerti}, D. and {Jansen}, J.~P.},
 eprint = {2102.09437},
 file = {:C\:/Users/Nathan/Documents/docs/2102.09437.pdf:pdf},
 keywords = {cost-effectiveness analysis,health economic evaluation,multi-state,simulation},
@@ -34,7 +34,7 @@ year = {2021}
 abstract = {Health economic evaluation studies are widely used in public health to assess health strategies in terms of their cost-effectiveness and inform public policies. We developed an R package for Markov models implementing most of the modelling and reporting features described in reference textbooks and guidelines: deterministic and probabilistic sensitivity analysis, heterogeneity analysis, time dependency on state-time and model-time (semi-Markov and non-homogeneous Markov models), etc. In this paper we illustrate the features of heemod by building and analysing an example Markov model. We then explain the design and the underlying implementation of the package.},
 archivePrefix = {arXiv},
 arxivId = {1702.03252},
-author = {Filipovi{\'{c}}-Pierucci, A., Zarca, K., & Durand-Zaleski, I.},
[…]
