### 2022-09-01 #4206 @n8thangreen
Hi, and thanks @ajstewartlang.

* I've edited the references following the example on the JOSS website so hopefully this is ok now
* I've tagged the latest release of the package here https://github.com/n8thangreen/BCEA/releases/tag/v2.4.2
* I've archived the package files on Zenodo here https://zenodo.org/record/7040450#.YxCDhXbMJD8
* The DOI is 10.5281/zenodo.7040450

I think this has covered all the points above. Please let me know if there's anything else I need to do. Thanks!
