### 2022-09-02 2a74aad
Update paper.bib
add doi for Baio (2013) - check this is the correct doi and date please. The entry [here](https://pubmed.ncbi.nlm.nih.gov/24343868/) indicates a 2014 publication data. If 2014 is correct, can you update paper.bib please?
