+++ paper.bib
@@ -47,6 +47,7 @@ year = {2017}
 
 @article{Baio:2013,
 	author = {{Baio}, G.},
+	doi = {10.1002/sim.6074},
 	title = {{Bayesian models for cost-effectiveness analysis in the presence of structural zero costs}},
 	journal = {Statistics in Medicine},
 	year = {2013},
