+++ paper.bib
@@ -45,11 +45,11 @@ volume = {1},
 year = {2017}
 }
 
-@article{Baio:2013,
+@article{Baio:2014,
 	author = {{Baio}, G.},
 	title = {{Bayesian models for cost-effectiveness analysis in the presence of structural zero costs}},
 	journal = {Statistics in Medicine},
-	year = {2013},
+	year = {2014},
 	volume = {33(11)},
 	pages = {1900-1913},
 }
