+++ paper.bib
@@ -52,6 +52,7 @@ year = {2017}
 	year = {2014},
 	volume = {33(11)},
 	pages = {1900-1913},
+	doi = {10.1002/sim.6074}
 }
 
 @article{Alarid-Escudero2019a,
