+++ paper.bib
@@ -14,7 +14,9 @@ author = {{Baio}, G. and {Berardi}, A. and {Heath}, A.},
 publisher = {Springer International Publishing},
 title = {{Bayesian Cost-Effectiveness Analysis with the R package BCEA}},
 url = {https://www.springer.com/gp/book/9783319557168},
-year = {2017}
+year = {2017},
+doi = {10.1007/978-3-319-55718-2},
+isbn = {978-3-319-55716-8}
 }
 
 @article{Incerti2021,
