+++ paper.bib
@@ -12,7 +12,7 @@
 abstract = {The book provides a description of the process of health economic evaluation and modelling for cost-effectiveness analysis, particularly from the perspective of a Bayesian statistical approach. Some relevant theory and introductory concepts are presented using practical examples and two running case studies. The book also describes in detail how to perform health economic evaluations using the R package BCEA (Bayesian Cost-Effectiveness Analysis). BCEA can be used to post-process the results of a Bayesian cost-effectiveness model and perform advanced analyses producing standardised and highly customisable outputs. It presents all the features of the package, including its many functions and their practical application, as well as its user-friendly web interface. The book is a valuable resource for statisticians and practitioners working in the field of health economics wanting to simplify and standardise their workflow, for example in the preparation of dossiers in support of marketing authorisation, or academic and scientific publications.},
 author = {{Baio}, G. and {Berardi}, A. and {Heath}, A.},
 publisher = {Springer International Publishing},
-title = {{Bayesian Cost-Effectiveness Analysis with the R package BCEA}},
+title = {Bayesian Cost-Effectiveness Analysis with the {R} package {BCEA}},
 url = {https://www.springer.com/gp/book/9783319557168},
 year = {2017},
 doi = {10.1007/978-3-319-55718-2},
@@ -27,7 +27,7 @@ author = {{Incerti}, D. and {Jansen}, J.~P.},
 eprint = {2102.09437},
 file = {:C\:/Users/Nathan/Documents/docs/2102.09437.pdf:pdf},
 keywords = {cost-effectiveness analysis,health economic evaluation,multi-state,simulation},
-title = {{hesim: Health Economic Simulation Modeling and Decision Analysis}},
+title = {{h}esim: Health Economic Simulation Modeling and Decision Analysis},
 url = {http://arxiv.org/abs/2102.09437},
 year = {2021}
 }
@@ -41,7 +41,7 @@ eprint = {1702.03252},
 file = {:C\:/Users/Nathan/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Filipovi{\'{c}}-Pierucci, Zarca, Durand-Zaleski - 2017 - Markov Models for Health Economic Evaluations The R Package heemod.pdf:pdf},
 keywords = {health economic evaluation,markov models,r},
 number = {1},
-title = {{Markov Models for Health Economic Evaluations: The R Package heemod}},
+title = {{M}arkov Models for Health Economic Evaluations: The {R} Package {heemod}},
 url = {http://arxiv.org/abs/1702.03252},
 volume = {1},
 year = {2017}
@@ -49,7 +49,7 @@ year = {2017}
 
 @article{Baio:2014,
 	author = {{Baio}, G.},
-	title = {{Bayesian models for cost-effectiveness analysis in the presence of structural zero costs}},
+	title = {{B}ayesian models for cost-effectiveness analysis in the presence of structural zero costs},
 	journal = {Statistics in Medicine},
 	year = {2014},
 	volume = {33(11)},
@@ -69,7 +69,7 @@ mendeley-groups = {r-hta website},
 number = {11},
 pages = {1329--1339},
 publisher = {Springer International Publishing},
-title = {{A Need for Change! A Coding Framework for Improving Transparency in Decision Modeling}},
+title = {A Need for Change! A Coding Framework for Improving Transparency in Decision Modeling},
 url = {https://doi.org/10.1007/s40273-019-00837-x},
 volume = {37},
 year = {2019}
@@ -87,7 +87,7 @@ keywords = {Markov model,R project,decision-analytic modeling,microsimulation,op
 number = {3},
[…]
