### 2022-09-12 #4206 @danielskatz
@n8thangreen - I'm the AEiC currently working on accepting and publishing your paper.  I'm suggesting some style changes in https://github.com/n8thangreen/BCEA/pull/121 and https://github.com/n8thangreen/BCEA/pull/122 - please merge these, or let me know what you disagree with, then we can proceed
